Fix: Use Pass dataclass for different acceptances.
The old version of this one accepted 40 hands of all the weird types,
but it was up to the dealer and probability to determine how many of
each were created.

Used the `Pass` dataclass and the `generate_pbn_passes` function to ensure there
are enough of each "weird" hand to practise with.